Refactor LocalResponseCacheGatewayFilterFactory to ResponseCacheGatewayFilter.  Split out LocalResponseCacheAutoConfiguration and RedisResponseCacheAutoConfiguration
