+++ spring-cloud-gateway-server/src/main/java/org/springframework/cloud/gateway/config/ConfigurableHintsRegistrationProcessor.java
@@ -38,7 +38,7 @@ import org.springframework.cloud.gateway.filter.factory.FallbackHeadersGatewayFi
 import org.springframework.cloud.gateway.filter.factory.JsonToGrpcGatewayFilterFactory;
 import org.springframework.cloud.gateway.filter.factory.SpringCloudCircuitBreakerResilience4JFilterFactory;
 import org.springframework.cloud.gateway.filter.factory.TokenRelayGatewayFilterFactory;
-import org.springframework.cloud.gateway.filter.factory.cache.LocalResponseCacheGatewayFilterFactory;
+import org.springframework.cloud.gateway.filter.factory.cache.ResponseCacheGatewayFilterFactory;
 import org.springframework.cloud.gateway.filter.ratelimit.RedisRateLimiter;
 import org.springframework.cloud.gateway.support.Configurable;
 import org.springframework.context.annotation.ClassPathScanningCandidateComponentProvider;
@@ -75,7 +75,7 @@ class ConfigurableHintsRegistrationProcessor implements BeanFactoryInitializatio
 					"org.springframework.web.reactive.DispatcherHandler"),
 			SpringCloudCircuitBreakerResilience4JFilterFactory.class, circuitBreakerConditionalClasses,
 			FallbackHeadersGatewayFilterFactory.class, circuitBreakerConditionalClasses,
-			LocalResponseCacheGatewayFilterFactory.class,
+			ResponseCacheGatewayFilterFactory.class,
 			Set.of("com.github.benmanes.caffeine.cache.Weigher", "com.github.benmanes.caffeine.cache.Caffeine",
 					"org.springframework.cache.caffeine.CaffeineCacheManager"));
